removed unused variables, added some returns, and replaced the moving average with simple moving average for dynamic resizing
